gpu: nvgpu: gp10b: misc MISRA 10.3 fixes
This fixes some miscellaneous MISRA 10.3 violations in gp10b for
assignment of objects of different size or essential type.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2006586
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/gp10b/fifo_gp10b.c
@@ -90,7 +90,7 @@ int channel_gp10b_setup_ramfc(struct channel_gk20a *c,
 
 	nvgpu_mem_wr32(g, mem, ram_fc_gp_base_hi_w(),
 		pbdma_gp_base_hi_offset_f(u64_hi32(gpfifo_base)) |
-		pbdma_gp_base_hi_limit2_f(ilog2(gpfifo_entries)));
+		pbdma_gp_base_hi_limit2_f((u32)ilog2(gpfifo_entries)));
 
 	nvgpu_mem_wr32(g, mem, ram_fc_signature_w(),
 		 c->g->ops.fifo.get_pbdma_signature(c->g));
